+++ models/CalendarEntry.php
@@ -103,7 +103,6 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
      * Filters
      */
     const FILTER_PARTICIPATE = 1;
-    const FILTER_INVITED = 2;
     const FILTER_NOT_RESPONDED = 3;
     const FILTER_RESPONDED = 4;
     const FILTER_MINE = 5;
@@ -259,8 +258,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
 
         $participants = $this->getParticipantUsersByState([
             CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE,
-            CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED,
-            CalendarEntryParticipant::PARTICIPATION_STATE_INVITED]);
+            CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED]);
 
         if($this->closed) {
             CanceledEvent::instance()->from(Yii::$app->user->getIdentity())->about($this)->sendBulk($participants);
@@ -273,8 +271,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
     {
         $participants = $this->getParticipantUsersByState([
             CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE,
-            CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED,
-            CalendarEntryParticipant::PARTICIPATION_STATE_INVITED]);
+            CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED]);
 
         EventUpdated::instance()->from(Yii::$app->user->getIdentity())->about($this)->sendBulk($participants);
     }
@@ -476,6 +473,34 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
         return false;
     }
 
+    public function setParticipationState($type, User $user = null) {
+        if ($user == null && !Yii::$app->user->isGuest) {
+            $user = Yii::$app->user->getIdentity();
+        }
+
+        // TODO return a calendarEntryParticipant with errors explaining why
+        if(!$this->canRespond()) {
+            return null;
+        }
+
+        $calendarEntryParticipant = $this->findParticipant($user);
+
+        if ($calendarEntryParticipant == null) {
+            $calendarEntryParticipant = new CalendarEntryParticipant([
+                'user_id' => $user->id,
+                'calendar_entry_id' => $this->id]);
+        }
+
+        if($type === CalendarEntryParticipant::PARTICIPATION_STATE_NONE) {
+            // never explicitly store PARTICIPATION_STATE 0
+            $calendarEntryParticipant->delete();
+        } else {
+            $calendarEntryParticipant->participation_state = $type;
+            $calendarEntryParticipant->save();
+        }
+        return $calendarEntryParticipant;
+    }
+
     public function getParticipationState(User $user = null)
     {
         if (Yii::$app->user->isGuest) {
@@ -493,7 +518,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
             return $participant->participation_state;
         }
 
-        return 0;
+        return CalendarEntryParticipant::PARTICIPATION_STATE_NONE;
     }
 
     /**
@@ -630,16 +655,17 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
     {
         $participant = $this->findParticipant();
 
-        $result = '';
-        if($participant) {
+        if($participant && $this->isParticipationAllowed()) {
             switch($participant->participation_state) {
                 case CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED:
                     return Label::success(Yii::t('CalendarModule.base', 'Attending'))->right();
-                case CalendarEntryParticipant::PARTICIPATION_STATE_INVITED:
-                    return Label::success(Yii::t('CalendarModule.base', 'Invited'))->right();
                 case CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE:
-                    return Label::success(Yii::t('CalendarModule.base', 'Interested'))->right();
+                    if($this->allow_maybe) {
+                        return Label::success(Yii::t('CalendarModule.base', 'Interested'))->right();
+                    }
             }
         }
+
+        return null;
     }
 }